+++ glfw/wl_platform.h
@@ -70,7 +70,6 @@ typedef VkBool32 (APIENTRY *PFN_vkGetPhysicalDeviceWaylandPresentationSupportKHR
 #include "wayland-xdg-toplevel-icon-v1-client-protocol.h"
 #include "wayland-xdg-system-bell-v1-client-protocol.h"
 #include "wayland-xdg-toplevel-tag-v1-client-protocol.h"
-#include "wayland-color-management-v1-client-protocol.h"
 
 #define _glfw_dlopen(name) dlopen(name, RTLD_LAZY | RTLD_LOCAL)
 #define _glfw_dlclose(handle) dlclose(handle)
@@ -172,7 +171,6 @@ typedef struct _GLFWwindowWayland
     bool                        hovered;
     bool                        transparent;
     struct wl_surface*          surface;
-    struct wp_color_management_surface_v1 *color_management;
     bool                        waiting_for_swap_to_commit;
     struct wl_egl_window*       native;
     struct wl_callback*         callback;
@@ -348,7 +346,6 @@ typedef struct _GLFWlibraryWayland
     struct xdg_toplevel_icon_manager_v1* xdg_toplevel_icon_manager_v1;
     struct xdg_system_bell_v1* xdg_system_bell_v1;
     struct xdg_toplevel_tag_manager_v1* xdg_toplevel_tag_manager_v1;
-    struct wp_color_manager_v1* wp_color_manager_v1;
     struct wp_cursor_shape_manager_v1* wp_cursor_shape_manager_v1;
     struct wp_cursor_shape_device_v1* wp_cursor_shape_device_v1;
     struct wp_fractional_scale_manager_v1 *wp_fractional_scale_manager_v1;
@@ -398,14 +395,6 @@ typedef struct _GLFWlibraryWayland
         PFN_wl_egl_window_destroy window_destroy;
         PFN_wl_egl_window_resize window_resize;
     } egl;
-
-    struct {
-        struct { bool gamma22, ext_linear, srgb; } supported_transfer_functions;
-        struct { bool srgb; } supported_primaries;
-        struct { bool parametric, set_primaries; } supported_features;
-        bool capabilities_reported, image_description_done, has_needed_capabilities;
-        struct wp_image_description_v1 *image_description;
-    } color_manager;
 
     struct {
         glfw_wl_xdg_activation_request *array;